Refactor og rydding av journalførings-item til listen.
