+++ src/felles-komponenter/forside/saksliste/journalForingEnkeltLinje.less
@@ -1,0 +1,33 @@
+@import '../../../utils/mixins';
+
+.journalEnkeltLinje {
+  margin: 0.5em 0;
+  flex: 1;
+}
+
+.journalEnkeltLinje__link:link,
+.journalEnkeltLinje__link:visited,
+.journalEnkeltLinje__link:hover {
+  color: @black;
+  text-decoration: none;
+  flex-grow:1;
+  display: flex;
+}
+
+.journalEnkeltLinje__meta {
+  display: block;
+  margin: 0.5em 0;
+}
+
+.journalEnkeltLinje__meta__term {
+  float: left;
+  clear: left;
+  font-weight: bold;
+  margin-right: 0.4em;
+}
+
+.journalEnkeltLinje__meta__detalj {
+  float:left;
+  clear: right;
+  margin: 0;
+}
