--- src/felles-komponenter/forside/saksliste/sakListe.less
@@ -1,10 +1,0 @@
-@import '../../../constants';
-
-.sokliste {
-  margin: 0 0 2em 0;
-
-  .sokeliste__fleresaker:visited,
-  .sokeliste__fleresaker:link {
-    color:@black;
-  }
-}
