--- src/felles-komponenter/sok/sokListeEnkeltlinje.less
@@ -1,65 +1,0 @@
-@import '../../constants';
-
-.sokliste__enkeltlinje {
-  display:flex;
-  flex-grow:1;
-  flex-direction: row;
-  margin:1em 0;
-
-  .enkeltlinje__panel {
-    display:flex;
-    flex-direction: row;
-    flex-grow:1;
-    border:3px solid transparent;
-    transition: all 0.5s;
-  }
-
-  &.sakErNy {
-    .enkeltlinje__panel {
-      animation: 1s linear slidein;
-      animation-iteration-count: 3;
-    }
-
-    @keyframes slidein {
-      0% {
-        border-color: transparent;
-      }
-
-      50% {
-        border-color: @orange;
-      }
-
-      100% {
-        border-color: transparent;
-      }
-    }
-  }
-
-  .enkeltlinje__link:link,
-  .enkeltlinje__link:visited,
-  .enkeltlinje__link:hover {
-    color: @black;
-    text-decoration: none;
-    flex-grow:1;
-    display: flex;
-  }
-
-  .enkeltlinje__kjoenn {
-    display: flex;
-    flex-basis:1em;
-    justify-content: center;
-    padding-right:1em;
-
-    .kjoenn__ikon {
-      width: 2em;
-      height: 2em;
-    }
-  }
-
-  .enkeltlinje__info {
-    display: flex;
-    flex-direction: column;
-    flex-grow: 5;
-    flex-basis:5em;
-  }
-}
